app/requirements.txt
Flask==2.1.2
Flask_SQLAlchemy==2.5.1
Flask_WTF==1.0.1
folium==0.12.1.post1
geopandas==0.11.1
gunicorn==20.1.0
ipython==8.4.0
scikit-learn==1.0.2
joblib==1.1.0
numpy==1.23.0
pandas==1.4.3
passlib==1.7.4
python-dotenv==0.20.0
vincent==0.4.4
WTForms==3.0.1
lxml==4.9.1
psycopg2==2.9.3

